Support for overriding type of a sequence component
This feature is motivated by X509 certificate extensions. Extension is defined in RFC5280 as a sequence type with components "extnID", "criticial" and "extnValue".
Component "extnValue" is of type OCTET STRING, but its value is actually a DER encoding of another type that is defined by OID value of "extnId" component.

Every ASN.1 module can now specify types that have such components ("Extension"). Every override is defined as a pair - name of the OID component ("extnID") and name of the value component ("extnValue"). With the correct definition in place, component "extnValue" will be interpreted as an ExplicitTag element instead of OctetString. Inside the ExplicitTag element, there will be an element of type defined by "extnID".